Commit unchanged sources from Boost 1.58

+++ libs/python/src/object/function.cpp
@@ -754,8 +754,7 @@ object function_object(
     return python::object(
         python::detail::new_non_null_reference(
             new function(
-                f, keywords.first, 
-                (unsigned) (keywords.second - keywords.first))));
+                f, keywords.first, keywords.second - keywords.first)));
 }
 
 object function_object(py_function const& f)
